+++ src/patterns/zora-tab-bar/index.ts
@@ -1,2 +1,0 @@
-export type { ZoraTabBarProps } from './types';
-export { ZoraTabBar } from './ZoraTabBar';